Misc fixes to Coremark build and SV linting.
Changed top Makefile to let users override TARGETS to a subset of their choice.

Added automated population of Coremark code. Fixed generating *.dump and *.hex
files for Coremark.

Started fixing Verilator linting issues on the RTL code.

Resolve conflicts caused by cherry-pick commit d4b80e4 from brabect1 fork. The change related to the addition of functionality that allows the user to add their own targets was skipped, due to the fact that it is already present in the build system. The rest of the changes from the commit were applied.